requirements.txt
numpy==1.9.2
boto==2.36.0
botocore==1.1.8
certifi==2015.4.28
colorama==0.3.3
decorator==3.4.0
docutils==0.12
Flask==0.10.1
funcsigs==0.4
itsdangerous==0.24
Jinja2==2.8
jmespath==0.7.1
jsonschema==2.4.0
librato-metrics==0.4.9
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.7
mixpanel-py==3.1.1
mock==1.3.0
multipledispatch==0.4.8
networkx==1.10
nltk==3.0.4
nose==1.3.7
pandas==0.16.2
pbr==1.4.0
Pillow==2.9.0
prettytable==0.7.2
psycopg2==2.6
ptyprocess==0.4
pyasn1==0.1.8
Pygments==2.0.2
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
pyzmq==14.7.0
requests==2.3.0
rsa==3.1.4
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.16.0
six==1.9.0
sseclient==0.0.8
terminado==0.5
tornado==4.1
virtualenv==13.0.1
Werkzeug==0.10.4
